Introduced smearing which depends on the parameterisation supplied by the LC-TPC group. It is parameterised in phi and theta. Not yet in z.
The Covariance matrix has been corrected so that it is set for XYZ and not r-phi as previously.
The corresponding conversion back to r-phi has been set in LEPTracking and FullLDCTracking.

Code also added to provide detailed histogramming, which can be enabled for complilation by setting the ifdef STEVESCHECKPLOTS 1 in tpcdigiproccessor.h